requirements.txt
Django>=3.1.2,<3.2.0
djangorestframework>=3.12.1,<3.13.0
flake8==3.8.4
psycopg2>=2.8.6,<2.9.0